apispec Convert player monitoring unit tests to integration tests.
